Explicitly wait for Geonames engine to report it is ready for autocomplete services
The Geonames engine is ready to serve most plugins much faster than
autocomplete. Hence the engine was changed to start a separate thread
for initializing autocomplete data, and a new method was added for
checking whether the autocomplete data is ready. Autocomplete now
polls the method in 100ms intervals before registering itself to the
server.